% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{conjugate-models}
\alias{conjugate-models}
\alias{beta_binomial}
\title{Bayesian Models of the Conjugate Distributions}
\usage{
beta_binomial(prior, y)
}
\arguments{
\item{prior}{A parameters list of the prior distribution.}

\item{y}{A vector (or matrix) of observed data.}
}
\value{
An object of class \code{conjugate}.
}
\description{
Compute components for the bayesian models of the
    conjugate distributions.
}
\author{
Eduardo Junior <edujrrib@gmail.com>

Eduardo Junior <edujrrib@gmail.com>
}
