% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{draw_legend}
\alias{draw_legend}
\title{Produce a Legend or Key}
\usage{
draw_legend(lines = TRUE, ...)
}
\arguments{
\item{lines, ...}{Parameter into a list to create key.}
}
\description{
Produce a legend or key based on a list of arguments.
}
