% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methods-conjugate}
\alias{methods-conjugate}
\alias{print.conjugate}
\alias{confint.conjugate}
\title{S3 Methods for the Object of Conjugate Class}
\usage{
\method{print}{conjugate}(x, ...)

\method{confint}{conjugate}(object, parm = "theta", level = 0.95,
  type = c("quantile", "hpd"), ...)
}
\arguments{
\item{x}{An object of class \code{conjugate}.}

\item{...}{unused.}

\item{object}{An object of class \code{conjugate}.}

\item{parm}{Unused.}

\item{level}{The confidence level required.}

\item{type}{The type of interval.}
}
\description{
Methods functions
}
\author{
Eduardo Junior <edujrrib@gmail.com>

Eduardo Junior <edujrrib@gmail.com>
}
