% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model2dist}
\alias{model2dist}
\title{Parser of the Model Name to Probability Distribution}
\usage{
model2dist(model)
}
\arguments{
\item{model}{An object of class \code{conjugate}.}
}
\value{
The [d, p, q, r] functions of the congate distribution.
}
\description{
Get the probability distribution functions based on
    conjugacy name.
}
\author{
Eduardo Junior <edujrrib@gmail.com>
}
